--- ActionsA/state_transitions.csv
@@ -1,1 +1,0 @@
-current_state,action_on,action_params,next_state
