--- ActionsA/state_user.csv
@@ -1,1 +1,0 @@
-user_id,current_state
