+++ core/natmodel.umdl
