+++ logs/logs.csv
@@ -44,3 +44,4 @@ uid,req,res,timestamp,date
 101,who is the weather in boston,The weather is Sunny reaching 40 degree celsius,1491887832016,04/11/2017
 101,what is the time,It is 14 38,1492420084730,04/17/2017
 101,how is the weather,The weather is Sunny reaching 41 degree celsius,1492420103477,04/17/2017
+101,what's the time,It is 10 7,1492663061429,04/20/2017
