+++ requirements.txt
@@ -10,3 +10,4 @@ vine=1.1.3
 pymongo=3.4.0
 pygame
 gtts
+tasktiger=0.8.4
